+++ modules/apps/bookmarks/bookmarks-service/src/main/java/com/liferay/bookmarks/internal/search/BookmarksFolderBatchReindexer.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.bookmarks.internal.search.util;
+package com.liferay.bookmarks.internal.search;
 
 /**
  * @author Luan Maoski
